Add files via upload
The ONLINE QUIZ AND CONTEST MANAGEMENT SYSTEM is a web
application for to take online quiz in an efficient manner and no time wasting
for checking the paper. The main objective of ONLINE QUIZ AND CONTEST
MANAGEMENT SYSTEM is to efficiently evaluate the candidate thoroughly
through a fully automated system that not only saves lot of time but also gives
fast results.